+++ ctdcal/processors/functions_salt.py
@@ -1,0 +1,39 @@
+"""
+Functions to convert data from original to engineering units, engineering
+to science units, and to produce derived variables.
+"""
+import numpy as np
+from gsw import SP_salinometer, C_from_SP
+
+
+def CR_to_cond(cr, bath_t, ref_t, btl_p):
+    """
+    Convert AutoSal double conductivity ratio (CR) to conductivity using
+    GSW conversion routines.
+
+    Parameters
+    ----------
+    cr : array-like
+        Double conductivity ratio from AutoSal, unitless
+    bath_t : array-like
+        AutoSal water bath temperature, degrees C
+    ref_t : array-like
+        CTD temperature at bottle stop, degrees C
+    btl_p : array-like
+        CTD pressure at bottle stop, dbar
+
+    Returns
+    -------
+    cond : array-like
+        Converted reference conductivity, mS/cm
+
+    """
+
+    salinity = SP_salinometer((cr / 2.0), bath_t)
+    cond = C_from_SP(salinity, ref_t, btl_p)
+
+    # ignore RunTimeWarning from (np.nan <= 1)
+    with np.errstate(invalid="ignore"):
+        cond[cond <= 1] = np.nan
+
+    return cond
