+++ ctdcal/processors/proc_bottle.py
@@ -1,0 +1,56 @@
+"""
+Transforming processed bottle data into derivative forms (binned ctd bottle data,
+merged water sample data)
+"""
+import logging
+from pathlib import Path
+
+import pandas as pd
+
+from ctdcal import get_ctdcal_config
+from ctdcal.process_bottle import retrieveBottleData, bottle_mean
+
+cfg = get_ctdcal_config()
+log = logging.getLogger(__name__)
+
+
+def make_btl_mean(ssscc_list):
+    """
+    Create "bottle mean" files from continuous CTD data averaged at the bottle stops.
+
+    Parameters
+    ----------
+    ssscc_list : list of str
+        List of stations to convert
+
+    Returns
+    -------
+    boolean
+        bottle averaging of mean has finished successfully
+    """
+    log.info("Generating btl_mean.pkl files")
+    for ssscc in ssscc_list:
+        if not Path(cfg.dirs["bottle"] + ssscc + "_btl_mean.pkl").exists():
+            imported_df = pd.read_pickle(cfg.dirs["converted"] + ssscc + ".pkl")
+            bottle_df = retrieveBottleData(imported_df)
+            mean_df = bottle_mean(bottle_df)
+
+            # export bottom bottle time/lat/lon info
+            fname = cfg.dirs["logs"] + "bottom_bottle_details.csv"
+            datetime_col = "nmea_datetime"
+            if datetime_col not in mean_df.columns:
+                log.debug(
+                    f"'{datetime_col}' not found in DataFrame - using 'scan_datetime'"
+                )
+                datetime_col = "scan_datetime"
+
+            bot_df = mean_df[[datetime_col, "GPSLAT", "GPSLON"]].head(1)
+            bot_df.columns = ["bottom_time", "latitude", "longitude"]
+            bot_df.insert(0, "SSSCC", ssscc)
+            add_header = not Path(fname).exists()  # add header iff file doesn't exist
+            with open(fname, "a") as f:
+                bot_df.to_csv(f, mode="a", header=add_header, index=False)
+
+            mean_df.to_pickle(cfg.dirs["bottle"] + ssscc + "_btl_mean.pkl")
+
+    return True
